@@ -5,10 +5,11 @@
     "keywords": ["nette", "rte", "typo3"],
     "license": "MIT",
     "require": {
-        "php": ">=5.3.3"
+        "php": ">=5.6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.3"
+        "phpunit/phpunit": "^5.5",
+        "squizlabs/php_codesniffer": "^2.0"
     },
     "autoload": {
         "classmap": ["src/"] 
